+++ packages/data/src/index.ts
@@ -1,4 +1,4 @@
-export { DataIntegrationTest, isDataIntegrationTest } from './data-integration-test.js';
+export { DataIntegrationTest, declareDataIntegrationTest, isDataIntegrationTest } from './data-integration-test.js';
 export type { DataIntegrationTestClass } from './data-integration-test.js';
 export type { DataIntegrationTestConfiguration } from './data-integration-test-configuration.js';
 export type { DataIntegrationTestContextAccessor } from './data-integration-test-context-accessor.js';
@@ -14,3 +14,4 @@ export {
   consoleDataIntegrationTestLogger,
 } from './logging/console-data-integration-test-logger.js';
 export type { DataIntegrationTestLogger } from './logging/data-integration-test-logger.js';
+export { createDataIntegrationTestContext } from './create-data-integration-test-context.js';